Refactored component and made custom hooks for text measurement and layout checking
