+++ src/components/AnswerToggle.tsx
@@ -1,8 +1,10 @@
-import { useEffect, useState, useRef, useCallback } from 'react';
-import AnswerToggleButton from "./AnswerToggleButton";
+import React from "react";
 import { motion } from "framer-motion";
+import AnswerToggleButton from "./AnswerToggleButton";
 import { OptionModel } from "../models";
 import { cn } from "../utils";
+import useTextMeasurement from "../hooks/useTextMeasurement";
+import useLayoutCheck from "../hooks/useLayoutCheck";
 
 interface AnswerToggleProps {
   options: OptionModel[];
@@ -11,62 +13,19 @@ interface AnswerToggleProps {
   onToggle: (selectedOption: OptionModel) => void;
 }
 
-const AnswerToggle = ({
+const AnswerToggle: React.FC<AnswerToggleProps> = ({
   options,
   selectedOption,
   isAllCorrect,
   onToggle,
-}: AnswerToggleProps) => {
-  const [isStacked, setIsStacked] = useState(false);
-  const containerRef = useRef<HTMLDivElement>(null);
-  const canvasRef = useRef<HTMLCanvasElement | null>(null);
+}) => {
+  const measureText = useTextMeasurement();
+  const { containerRef, isStacked } = useLayoutCheck(options, measureText);
 
-  const measureText = useCallback((text: string, font: string) => {
-    if (!canvasRef.current) {
-      canvasRef.current = document.createElement('canvas');
-    }
-    const context = canvasRef.current.getContext('2d');
-    if (context) {
-      context.font = font;
-      return context.measureText(text).width;
-    }
-    return 0;
-  }, []);
-
-  const checkLayout = useCallback(() => {
-    if (containerRef.current) {
-      const containerWidth = containerRef.current.offsetWidth;
-      const buttonWidth = containerWidth / options.length;
-      
-      const buttonStyle = window.getComputedStyle(containerRef.current.querySelector('button') as HTMLElement);
-      const font = `${buttonStyle.fontWeight} ${buttonStyle.fontSize} ${buttonStyle.fontFamily}`;
-
-      const shouldStack = options.some(option => {
-        const words = option.optionText.split(' ');
-        return words.some(word => {
-          const wordWidth = measureText(word, font);
-          return wordWidth > buttonWidth - 40; // Subtracting 40px for padding
-        });
-      });
-
-      setIsStacked(shouldStack);
-    }
-  }, [options, measureText]);
-
-  useEffect(() => {
-    checkLayout();
-    window.addEventListener('resize', checkLayout);
-
-    return () => {
-      window.removeEventListener('resize', checkLayout);
-    };
-  }, [checkLayout]);
-
-  // Safely determine if the first option is selected
   const isFirstOptionSelected = selectedOption ? selectedOption.id === options[0].id : true;
 
   return (
-    <div 
+    <div
       ref={containerRef}
       className={cn(
         "relative w-full max-w-screen-lg border-2 border-white border-opacity-60 md:rounded-full rounded-3xl overflow-hidden"
@@ -74,9 +33,10 @@ const AnswerToggle = ({
     >
       <motion.div
         className={cn(
-          "absolute bg-white bg-opacity-60 md:rounded-full rounded-2xl",
-          isStacked ? "w-full" : "w-1/2",
-          isStacked ? "h-1/2" : "h-full"
+          "absolute bg-white bg-opacity-60",
+          isStacked
+            ? "w-full h-1/2 rounded-none"
+            : "w-1/2 h-full md:rounded-full rounded-2xl"
         )}
         animate={{
           y: isStacked ? (isFirstOptionSelected ? "0%" : "100%") : 0,
@@ -84,7 +44,7 @@ const AnswerToggle = ({
         }}
         transition={{ type: "spring", stiffness: 300, damping: 30 }}
       />
-      <div 
+      <div
         className={cn(
           "relative z-10",
           isStacked ? "grid grid-rows-2 h-full" : "flex flex-row"
